Bugfix
Changed how evolutions are checked at the end of battles and how the data is stored at the start of battle. Now, if a Pokemon is caught and replaces a Pokemon in the player's party, only that newly caught Pokemon will check if it should evolve. Previously, if the player had that Pokemon in their party, unevolved, yet ready to, both the caught Pokemon and the Pokemon in the party would evolve sequentially. This is because the party indexes changed in GameplayUtils.rb::swapCaughtPokemon, which no longer happens as both Pokemon instances directly swap data.