Bugfix: Future Sight / Doom Desire
Resolved a major bug where neither Future Sight or Doom Desire would activate if the Pokemon which used it was fainted. Both now also calculate damage using the Pokemon's ability and held item. Certain abilities should not apply and should be handled separately in the future. ie. Beads of Ruin.