Feature: Trainer Level Scaling
The first instance of an entirely new feature where trainer Pokemon levels are scaled based on the player's Pokemon levels. All Pokemon are also force evolved based on a few conditions. This will be further modified, possibly refactored, and possibly optimized in the future.
There will also be another feature included where the moves of each Pokemon will be automatically selected based on various conditions; including, but not limited to, the Pokemon's normalized stats.